# back/ManualMovementsManager/ManualMovementsManager/src/ManualMovementsManager.Application/Commands/Customers/AddCustomer/AddCustomerResponse.cs
using ManualMovementsManager.Domain.Entities;
using ManualMovementsManager.Domain.Entities.Response;
using ManualMovementsManager.Application.DTOs;

namespace ManualMovementsManager.Application.Commands.Customers.AddCustomer
{
    public class AddCustomerResponse: BaseResponse<CustomerDto>
    {
    }
}